requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.11.18
aiohttp-retry==2.9.1
aiosignal==1.3.2
attrs==25.3.0
blinker==1.9.0
cachetools==5.5.2
certifi==2025.4.26
charset-normalizer==3.4.2
click==8.2.0
colorama==0.4.6
dotenv==0.9.9
Flask==3.1.0
flask-cors==6.0.0
frozenlist==1.6.0
google-auth==2.40.1
google-auth-oauthlib==1.2.2
gspread==6.2.1
gunicorn==23.0.0
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
multidict==6.4.3
oauthlib==3.2.2
packaging==25.0
propcache==0.3.1
pyasn1==0.6.1
pyasn1_modules==0.4.2
PyJWT==2.10.1
python-dotenv==1.1.0
requests==2.32.3
requests-oauthlib==2.0.0
rsa==4.9.1
twilio==9.6.0
urllib3==2.4.0
Werkzeug==3.1.3
yarl==1.20.0
